Sheet Component
Created a new component called Sheet. This component will be used
to add new inventory items. The idea is to more the NewMedia component
into it. This component is in its infancy with only a bit of logic and
styles. I also added a new add icon for the button in the component and
updated Dashboard component to include a window.confirm() just stop
accidental deletion of items during development. I'll later refine this
when I create my own Dialog component.

+++ src/views/Dashboard.vue
@@ -75,6 +75,7 @@
             </nas-modwrap>
         </section>
     </div>
+    <nas-sheet></nas-sheet>
 </template>
 
 <script>
@@ -136,17 +137,18 @@ export default {
     },
     methods: {
         async deleteMedia(event) {
-            console.log(`Received event from GC: ${event}`);
-            const response = await fetch(process.env.VUE_APP_API_URL, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify(event)
-            });
+            if (window.confirm('Do you really want to delete item?')) {
+                const response = await fetch(process.env.VUE_APP_API_URL, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify(event)
+                });
 
-            const data = await response.json();
-            console.log(`Server response: ${JSON.stringify(data)}`);
+                const data = await response.json();
+                console.log(`Server response: ${JSON.stringify(data)}`);
+            }
         },
         async getMedia() {
             const response = await fetch(process.env.VUE_APP_API_URL);
